fixing cache block error
git-svn-id: http://svn.concrete5.org/svn/concrete5@424 b0551a0c-1e16-4222-a7d5-975db1aca215

+++ trunk/web/concrete/models/collection.php
@@ -378,10 +378,10 @@ defined('C5_EXECUTE') or die(_("Access Denied."));
 			$vo = $this->getVersionObject();
 
 			Cache::delete('page', $this->getCollectionID());
-			Cache::delete('page', $this->getCollectionID() . ':' . $vo->getVersionID());
-
-			$v = $this->getVersionObject();
-			$v->refreshCache();
+			if (is_object($vo)) {
+				Cache::delete('page', $this->getCollectionID() . ':' . $vo->getVersionID());
+				$vo->refreshCache();
+			}
 		}
 		
 		public function getBlocks($arHandle = false) {
